+++ haskell-overlays/reflex-packages/default.nix
@@ -94,6 +94,14 @@ in
   chrome-test-utils = self.callCabal2nix "chrome-test-utils" (reflexDomRepo + "/chrome-test-utils") {};
 
   ##
+  ## Terminal / Conventional OS
+  ##
+
+  reflex-vty = self.callHackage "reflex-vty" "0.1.3.0" {};
+  reflex-process = self.callHackage "reflex-process" "0.1.0.1" {};
+  reflex-fsnotify = self.callHackage "reflex-fsnotify" "0.1.0.0" {};
+
+  ##
   ## GHCJS and JSaddle
   ##